- Added rendering upscaling to to have a small frame buffer for a larger window - Fixed MIS to work only for light sampling, use BSDF sampling for reflection - Fixed HDR textures for OpenGL - Created correct linear texture reading - Fixed issue with zero area triangles

+++ AwesomeRenderer/raytracer.h
@@ -36,7 +36,7 @@ namespace AwesomeRenderer
 			friend class RenderJob;
 
 		private:
-			static const float MAX_FRAME_TIME;
+			static const uint32_t MAX_FRAME_TIME;
 			static const uint32_t TILE_SIZE;
 
 			Timer frameTimer;
